requirements.txt
asgiref==3.6.0
Django==3.2.18
django-mptt==0.14.0
mysqlclient==2.1.1
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
mysqlclient==2.1.1
django-debug-toolbar==4.0.0
django-ckeditor==6.5.1